Refactor the navigation bar
