+++ Components/MainSection/MainSection.js
@@ -1,5 +1,4 @@
 import { Section as BaseSection } from '../Section'
-import { HomeNavigationBar } from '../NavigationBar/HomeNavigationBar'
 import { Alert } from '../Alert'
 import React from 'react'
 import styled from 'styled-components'
@@ -11,8 +10,6 @@ const Section = styled(BaseSection)`
 export const MainSection = () => {
   return (
     <Section column>
-      <HomeNavigationBar />
-
       <Alert isVisible top='70%' left='5%' />
     </Section>
   )
