+++ pages/experiences.js
@@ -2,7 +2,7 @@ import { ThemeProvider } from 'styled-components'
 import React from 'react'
 import { Footer } from '../Components/Footer'
 import { ExperienceSection } from '../Components/ExperienceSection'
-import { PagesNavigationBar } from '../Components/NavigationBar'
+import { NavigationBar } from '../Components/NavigationBar'
 
 // rem base is 16px.
 const theme = {
@@ -29,7 +29,7 @@ const theme = {
 const Experiences = () => (
   <ThemeProvider theme={theme}>
     <>
-      <PagesNavigationBar />
+      <NavigationBar />
 
       <ExperienceSection />
