+++ pages/index.js
@@ -1,4 +1,5 @@
 import { NextHead } from '../Components/Head'
+import { NavigationBar } from '../Components/NavigationBar'
 import { MainSection } from '../Components/MainSection'
 import { AboutSection } from '../Components/AboutSection'
 import { FacetsSection } from '../Components/FacetsSection'
@@ -34,6 +35,8 @@ const Home = () => (
   <ThemeProvider theme={theme}>
     <>
       <NextHead />
+
+      <NavigationBar isHome />
 
       <MainSection />
